Fix phase error when forward-reference used in FCB or FCW.
Forward-reference labels generate an error in pass 1 that should be ignored. But the logic must generate the same number of bytes of object code as it will do in pass 2, so that addresses match up. Probably useful as a check to add some code to compare label addresses in pass 2 with those recorded in pass 1 -- they should always be the same.