Implement RMB directive.
Reserve as many bytes as requested. Do not generate object code to fill the gap. Acts as a relative ORG (ORG .+NBYTES). However, unlike ORG, RMB is normally labelled to give the address of the reserved block. This works, but shows the wrong address in the listing file. The address in the symbol table is correct. Fixing this is a bigger problem, and would also allow ORG to be labelled (if that was useful).